feat: upgrade to spring boot 4

+++ api/common/src/main/java/dev/ngb/base_hub/common/infra/jdbc/entity/organization/OrganizationEntity.java
@@ -21,7 +21,7 @@ public record OrganizationEntity(
         OrganizationStatus status,
         Map<String, Object> configuration,
         @Version
-        Long version,
+        Integer version,
         String createdById,
         String updatedById,
         String deletedById,
